feat: improve archived list tile flow
