+++ lib/app/features/chat/recent_chats/views/components/conversation_edit_bottom_bar/components/conversation_archive_button/conversation_unarchive_button.dart
@@ -6,6 +6,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/extensions/extensions.dart';
 import 'package:ion/app/features/chat/recent_chats/providers/conversations_edit_mode_provider.c.dart';
 import 'package:ion/app/features/chat/recent_chats/providers/selected_conversations_ids_provider.c.dart';
+import 'package:ion/app/features/chat/recent_chats/providers/toggle_archive_conversation_provider.c.dart';
 import 'package:ion/generated/assets.gen.dart';
 
 class ConversationUnarchiveButton extends ConsumerWidget {
@@ -21,9 +22,9 @@ class ConversationUnarchiveButton extends ConsumerWidget {
       behavior: HitTestBehavior.opaque,
       onTap: () async {
         if (selectedConversations.isNotEmpty) {
-          // await ref
-          //     .read(e2eeConversationManagementProvider.notifier)
-          //     .toggleArchivedConversations(selectedConversations);
+          await ref
+              .read(toggleArchivedConversationsProvider.notifier)
+              .toogleConversation(selectedConversations);
           ref.read(conversationsEditModeProvider.notifier).editMode = false;
           ref.read(selectedConversationsProvider.notifier).clear();
           if (context.mounted && context.canPop()) {
