+++ lib/app/features/chat/recent_chats/views/pages/recent_chats_timeline_page/recent_chats_timeline_page.dart
@@ -29,7 +29,7 @@ class RecentChatsTimelinePage extends HookConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     useOnInit(() {
-      ref.read(toggleArchivedConversationsProvider);
+      ref.read(archivedConversationsProvider.notifier);
     });
 
     return CustomScrollView(
